디자인 이상한거 수정
