+++ app/src/main/java/womenproject/com/mybury/presentation/intro/SplashActivity.kt
@@ -1,5 +1,6 @@
 package womenproject.com.mybury.presentation.intro
 
+import android.annotation.SuppressLint
 import android.content.Intent
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
@@ -8,13 +9,19 @@ import womenproject.com.mybury.R
 import kotlin.random.Random
 import android.os.Handler
 import android.util.Log
+import io.reactivex.android.schedulers.AndroidSchedulers
+import io.reactivex.schedulers.Schedulers
+import womenproject.com.mybury.data.Preference
+import womenproject.com.mybury.data.UseUserIdRequest
+import womenproject.com.mybury.data.network.apiInterface
 import womenproject.com.mybury.databinding.SplashWithLoginBinding
+import womenproject.com.mybury.presentation.CanNotGoMainDialog
 import womenproject.com.mybury.presentation.MainActivity
 
 
 class SplashActivity : AppCompatActivity() {
 
-    lateinit var binding : SplashWithLoginBinding
+    lateinit var binding: SplashWithLoginBinding
     override fun onCreate(savedInstanceState: Bundle?) {
 
         super.onCreate(savedInstanceState)
@@ -32,10 +39,38 @@ class SplashActivity : AppCompatActivity() {
 
     private inner class splashhandler : Runnable {
         override fun run() {
-            val intent = Intent(context, MainActivity::class.java)
-            intent.flags = Intent.FLAG_ACTIVITY_NEW_TASK
-            context.startActivity(intent)
-            finish()
+            initToken()
         }
+    }
+
+    private fun goToNext() {
+        val intent = Intent(context, MainActivity::class.java)
+        intent.flags = Intent.FLAG_ACTIVITY_NEW_TASK
+        context.startActivity(intent)
+        finish()
+    }
+
+    @SuppressLint("CheckResult")
+    private fun initToken() {
+
+        val getTokenRequest = UseUserIdRequest(Preference.getUserId(this))
+
+        apiInterface.getLoginToken(getTokenRequest)
+                .subscribeOn(Schedulers.io())
+                .observeOn(AndroidSchedulers.mainThread())
+                .subscribe({ response ->
+                    if (response.retcode == "200") {
+                        Preference.setAccessToken(this, response.accessToken)
+                        Preference.setRefreshToken(this, response.refreshToken)
+                        goToNext()
+                    } else {
+                        CanNotGoMainDialog().show(supportFragmentManager, "tag")
+                    }
+                }) {
+                    Log.e("ayhan", it.toString())
+                    CanNotGoMainDialog().show(supportFragmentManager, "tag")
+                }
+
+
     }
 }
