+++ app/src/main/res/layout/bucket_item_count.xml
@@ -1,7 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
 
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools">
 
     <data>
 
@@ -43,19 +44,19 @@
 
                 <RelativeLayout
                     android:layout_width="match_parent"
-                    android:layout_height="match_parent"
+                    android:layout_height="wrap_content"
                     android:background="@drawable/mypage_category_background">
 
                     <ImageView
                         android:id="@+id/bucket_succeed_image"
                         android:layout_width="match_parent"
-                        android:layout_height="match_parent"
+                        android:layout_height="wrap_content"
                         android:background="@drawable/bucket_item_succeed_background" />
 
                     <ImageView
                         android:id="@+id/bucket_item_image"
                         android:layout_width="match_parent"
-                        android:layout_height="match_parent"
+                        android:layout_height="wrap_content"
                         android:layout_alignParentStart="true"
                         android:layout_alignParentTop="true"
                         android:background="@drawable/bucket_item_base_background" />
@@ -63,7 +64,7 @@
                     <RelativeLayout
                         android:id="@+id/count_succeed_layout"
                         android:layout_width="match_parent"
-                        android:layout_height="match_parent"
+                        android:layout_height="wrap_content"
                         android:layout_marginStart="16dp"
                         android:visibility="gone">
 
@@ -79,58 +80,68 @@
                     <RelativeLayout
                         android:id="@+id/content_layout"
                         android:layout_width="match_parent"
-                        android:layout_height="match_parent"
+                        android:layout_height="wrap_content"
                         android:layout_marginStart="16dp"
-                        android:layout_marginBottom="19dp"
                         android:gravity="center_vertical">
 
                         <LinearLayout
-                            android:id="@+id/prgress_layout"
-                            android:layout_width="wrap_content"
-                            android:layout_height="wrap_content"
-                            android:layout_below="@id/bucket_title"
-                            android:layout_alignStart="@id/bucket_title"
-                            android:layout_marginTop="10dp"
-                            android:gravity="center_vertical"
-                            android:orientation="horizontal">
-
-                            <ProgressBar
-                                android:id="@+id/horizontal_progress_bar"
-                                style="@android:style/Widget.ProgressBar.Horizontal"
-                                android:layout_width="180dp"
-                                android:layout_height="6dp"
-                                android:max="10"
-                                android:progress="10"
-                                android:progressDrawable="@drawable/base_horizontal_progressbar" />
+                            android:layout_centerVertical="true"
+                            android:orientation="vertical"
+                            android:layout_width="match_parent"
+                            android:layout_height="wrap_content">
 
                             <TextView
-                                android:id="@+id/user_count"
-                                android:layout_width="30dp"
-                                android:layout_height="14dp"
-                                android:layout_marginLeft="4dp"
-                                android:gravity="center"
-                                android:text="@{currentUserCount}"
-                                android:textColor="@color/_a6c6ff"
-                                android:textSize="11sp" />
+                                android:id="@+id/bucket_title"
+                                style="@style/BucketItemTitleTextStyle"
+                                android:layout_width="match_parent"
+                                android:layout_height="wrap_content"
+                                android:layout_centerVertical="true"
+                                android:layout_marginTop="21dp"
+                                android:layout_marginRight="70dp"
+                                android:maxLines="2"
+                                android:text="@{bucketTitleText}" />
+
+
+                            <LinearLayout
+                                android:id="@+id/prgress_layout"
+                                android:layout_width="wrap_content"
+                                android:layout_height="wrap_content"
+                                android:layout_below="@id/bucket_title"
+                                android:layout_alignStart="@id/bucket_title"
+                                android:layout_marginTop="10dp"
+                                android:gravity="center_vertical"
+                                android:orientation="horizontal"
+                                android:paddingBottom="19dp">
+
+                                <ProgressBar
+                                    android:id="@+id/horizontal_progress_bar"
+                                    style="@android:style/Widget.ProgressBar.Horizontal"
+                                    android:layout_width="180dp"
+                                    android:layout_height="6dp"
+                                    android:max="10"
+                                    android:progress="10"
+                                    android:progressDrawable="@drawable/base_horizontal_progressbar" />
+
+                                <TextView
+                                    android:id="@+id/user_count"
+                                    android:layout_width="wrap_content"
+                                    android:layout_height="wrap_content"
+                                    android:layout_marginLeft="4dp"
+                                    android:gravity="center"
+                                    android:text="@{currentUserCount}"
+                                    android:textColor="@color/_a6c6ff"
+                                    android:textSize="11sp"
+                                    tools:text='10' />
+
+                            </LinearLayout>
+
 
                         </LinearLayout>
-
-                        <TextView
-                            android:id="@+id/bucket_title"
-                            style="@style/BucketItemTitleTextStyle"
-                            android:layout_width="match_parent"
-                            android:layout_height="wrap_content"
-                            android:layout_centerVertical="true"
-                            android:layout_marginTop="21dp"
-                            android:layout_marginRight="70dp"
-                            android:maxLines="2"
-                            android:text="@{bucketTitleText}" />
-
                     </RelativeLayout>
 
                     <RelativeLayout
                         android:layout_width="64dp"
-                        android:layout_height="match_parent"
+                        android:layout_height="wrap_content"
                         android:layout_alignParentRight="true"
                         android:layout_centerInParent="true"
                         android:onClick="@{bucketSuccessClickListener}">
