+++ app/src/main/res/layout/dday_bucket_list.xml
@@ -23,18 +23,23 @@
         android:orientation="vertical">
 
         <LinearLayout
-            android:layout_width="42dp"
-            android:layout_height="20dp"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
             android:layout_marginLeft="12dp"
             android:background="@drawable/current_dday_badge">
 
             <TextView
-                android:layout_width="match_parent"
-                android:layout_height="match_parent"
-                android:text="@{ddayText}"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
                 android:gravity="center"
+                android:paddingStart="4dp"
+                android:paddingTop="2dp"
+                android:paddingEnd="4dp"
+                android:paddingBottom="2dp"
+                android:text="@{ddayText}"
                 android:textColor="@color/_ffffff"
-                android:textSize="12sp" />
+                android:textSize="12sp"
+                tools:text="D-10" />
         </LinearLayout>
