+++ app/src/main/res/layout/mypage_category_item.xml
@@ -21,20 +21,25 @@
 
 
     <RelativeLayout
-        android:onClick="@{itemOnClickListener}"
         android:layout_width="match_parent"
-        android:layout_height="50dp"
+        android:layout_height="wrap_content"
+        android:layout_marginBottom="8dp"
         android:background="@drawable/mypage_category_background"
-        android:layout_marginBottom="8dp">
+        android:gravity="center_vertical"
+        android:minHeight="50dp"
+        android:onClick="@{itemOnClickListener}">
 
         <TextView
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_centerVertical="true"
             android:layout_marginLeft="15dp"
+            android:paddingTop="14dp"
+            android:paddingBottom="14dp"
             android:text="@{category_name}"
             android:textColor="@color/_434343"
-            android:textSize="14sp" />
+            android:textSize="14sp"
+            tools:text="카테고리 이름" />
 
         <TextView
             android:layout_width="wrap_content"
@@ -43,7 +48,8 @@
             android:layout_centerVertical="true"
             android:layout_marginRight="15dp"
             android:text="@{categoryCount}"
-            android:textColor="@color/_5a95ff" />
+            android:textColor="@color/_5a95ff"
+            tools:text="10" />
 
 
     </RelativeLayout>
