+++ app/src/main/res/layout/mypage_scroll_layout.xml
@@ -35,9 +35,12 @@
             <RelativeLayout
                 android:id="@+id/dday_layout"
                 android:layout_width="match_parent"
-                android:layout_height="50dp"
+                android:layout_height="wrap_content"
+                android:layout_gravity="center_vertical"
                 android:layout_marginTop="12dp"
                 android:background="@drawable/mypage_ddaylist_background"
+                android:gravity="center_vertical"
+                android:minHeight="50dp"
                 android:onClick="@{ddayListClickListener}">
 
                 <TextView
@@ -45,6 +48,8 @@
                     android:layout_height="wrap_content"
                     android:layout_centerVertical="true"
                     android:layout_marginLeft="15dp"
+                    android:paddingTop="14dp"
+                    android:paddingBottom="14dp"
                     android:text="D-day 버킷리스트"
                     android:textColor="@color/_434343"
                     android:textSize="14sp" />
@@ -56,7 +61,8 @@
                     android:layout_alignParentRight="true"
                     android:layout_centerVertical="true"
                     android:layout_marginRight="15dp"
-                    android:textColor="@color/_ffca5a" />
+                    android:textColor="@color/_ffca5a"
+                    tools:text="10" />
 
 
             </RelativeLayout>
@@ -64,7 +70,7 @@
 
             <RelativeLayout
                 android:layout_width="match_parent"
-                android:layout_height="18dp"
+                android:layout_height="wrap_content"
                 android:layout_marginTop="16dp">
 
                 <TextView
@@ -75,12 +81,16 @@
                     android:textSize="12sp" />
 
                 <TextView
-                    android:layout_width="40dp"
+                    android:layout_width="wrap_content"
                     android:layout_height="match_parent"
                     android:layout_alignParentRight="true"
                     android:background="@drawable/mypage_category_edit_background"
                     android:gravity="center"
                     android:onClick="@{categoryEditClickListener}"
+                    android:paddingStart="6dp"
+                    android:paddingTop="2dp"
+                    android:paddingEnd="6dp"
+                    android:paddingBottom="2dp"
                     android:text="편집"
                     android:textColor="@color/_5a95ff"
                     android:textSize="11sp" />
