+++ app/src/main/res/layout/splash_blue.xml
@@ -10,6 +10,8 @@
     <TextView
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
+        android:layout_marginStart="64dp"
+        android:layout_marginEnd="64dp"
         android:layout_marginBottom="16dp"
         android:letterSpacing="-0.05"
         android:text="당신의 버킷리스트를 응원합니다."
@@ -28,12 +30,12 @@
         android:layout_marginEnd="64dp"
         android:background="@drawable/splash_blue_background"
         android:gravity="center"
+        android:paddingTop="6dp"
+        android:paddingBottom="6dp"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        android:paddingTop="6dp"
-        android:paddingBottom="6dp">
+        app:layout_constraintTop_toTopOf="parent">
 
 
         <ImageView
@@ -45,10 +47,10 @@
             android:background="@drawable/splash_logo_blue" />
 
         <ImageView
-            android:layout_marginEnd="7dp"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_alignParentEnd="true"
+            android:layout_marginEnd="7dp"
             android:background="@drawable/splash_check_blue" />
