+++ app/src/main/res/values/styles.xml
@@ -115,8 +115,8 @@
     </style>
 
     <style name="writeTopTextStyle">
-        <item name="android:layout_width">40dp</item>
-        <item name="android:layout_height">40dp</item>
+        <item name="android:layout_width">wrap_content</item>
+        <item name="android:layout_height">wrap_content</item>
         <item name="android:gravity">center</item>
         <item name="android:textColor">@color/register_text_color</item>
         <item name="android:textSize">16sp</item>
