+++ app/src/main/res/xml/my_page_header.xml
@@ -81,10 +81,11 @@
             <Constraint
                 android:id="@id/name"
                 android:layout_width="match_parent"
-                android:layout_height="24dp"
+                android:layout_height="wrap_content"
                 android:layout_marginStart="8dp"
                 android:layout_marginTop="8dp"
                 android:layout_marginEnd="8dp"
+                android:layout_marginBottom="14dp"
                 android:text="울고싶어요"
                 android:textColor="@color/_434343"
                 android:textSize="14sp"
